[hw,flash_ctrl] addr-xor directly after scrambling
To add a correlation between data and address, the data is xored
with the address. So far, this xor was done for the 32b bus output
of the flash_phy_rd module (after the read-buffer).
This commit moves it ahead and applies the xor directly after
descrambling.
In addition, the 7b ECC is computed before the data is stored in the
read-buffer. The integrity check is still performed at the output data.
(The read-buffer increases by 2*7*(ECC) flops per entry)

Signed-off-by: Michael Gautschi <mgautschi@lowrisc.org>